Two 1.0 gates, and the suite: the settled naming rules become a lint, Scripts/check-api-names.sh reading every public declaration under API/ against sixteen of them on every preflight (limits, magnitudes, factories, Boolean spellings, advance against step, resource loaders, Int seeds, highlight exponents, the hooks, the draw verbs, the state stack, the value vocabulary, String destinations, the lifecycle pairs, knob), the exceptions and their reasons in an allow file whose entries fail when they match nothing; its first run found fifty-three drifts, all renamed here without shims and named in the changelog: ten minimum/maximum limits, twenty-four UInt64 seeds now Int so seed: variation needs no cast, seven bare-adjective Booleans, five URL-only writers given a String path form, the physics snapshot loader spelling withExtension:, two generator intensities that were amounts, a labeled timeline step, the ping-pong pair's advance() now swap(), and the Syphon client's isActive and stop() now isConnected and disconnect(). The snippet gate reads the reference pages as a second tree, every Swift block under Docs/ compiled beside the Guide's with its own known-gaps list and a --record that keeps the pages it did not check, the wrapper now lifting imports from under a header comment and reading an access level before the type keyword; its first pass caught seven stale blocks (a ComposeLayer scale, two drawCircle calls without their labels, a drawPoint without at:, an Int text size, a Meander parameter under an old name, and an image loader the generator itself spelled as a function that does not exist) and one stale ollin docs anchor. The two runner-only failures on the last commit are fixed: the Vision flow sign test refuses itself where Vision has no compute device, and the frame-ring test reads the ring after every draw of a burst instead of once after eight. The sharded suite was green before the renames and after them.

+++ Examples/3D/Geometry/MeshGrowth/Sketch.swift
@@ -71,7 +71,7 @@ final class MeshGrowthDemo: Sketch {
         switch driver {
         case .chemistry:
             let growth = MeshGrowth(mesh: seedMesh, driver: .chemical(.coral),
-                                    edgeLength: coarseness, seed: UInt64(variation))
+                                    edgeLength: coarseness, seed: variation)
             // The chemistry's pattern covers only part of the surface where the
             // other drivers push everywhere, so it earns a faster rate; the
             // settle hands it a formed pattern to grow from on the first step.
@@ -81,7 +81,7 @@ final class MeshGrowthDemo: Sketch {
 
         case .curvature:
             return MeshGrowth(mesh: seedMesh, driver: .curvature,
-                              edgeLength: coarseness, seed: UInt64(variation))
+                              edgeLength: coarseness, seed: variation)
 
         case .rim:
             // Grow only near the equator. The band is soft-edged so the folds
@@ -90,13 +90,13 @@ final class MeshGrowthDemo: Sketch {
                                     driver: .field { position, _ in
                                         1 - smoothstep(0.05, 0.45, abs(position.y))
                                     },
-                                    edgeLength: coarseness, seed: UInt64(variation))
+                                    edgeLength: coarseness, seed: variation)
             growth.growthRate = 0.9
             return growth
 
         case .even:
             return MeshGrowth(mesh: seedMesh, driver: .uniform,
-                              edgeLength: coarseness, seed: UInt64(variation))
+                              edgeLength: coarseness, seed: variation)
         }
     }
 }
